requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
aiosqlite==0.20.0
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
httpx==0.27.2